+++ custom/bihar/reports/indicators/indicators.py
@@ -1,5 +1,6 @@
 from inspect import ismethod
 from dimagi.utils.decorators.memoized import memoized
+from dimagi.utils.html import format_html
 from custom.bihar.calculations.types import DoneDueCalculator, TotalCalculator
 from custom.bihar.models import CareBiharFluff
 from custom.bihar.utils import get_all_owner_ids_from_group
@@ -352,12 +353,29 @@ class IndicatorDataProvider(object):
         num, denom = map(sum, zip((0, 0), *pairs()))
         return num, denom
 
+    def get_indicator_value(self, indicator):
+        return "%s/%s" % self.get_indicator_data(indicator)
+
     @memoized
     def get_case_ids(self, indicator):
         return indicator.fluff_calculator.aggregate_results(
             ([self.domain, owner_id] for owner_id in self.all_owner_ids),
             reduce=False
         )[indicator.fluff_calculator.primary]
+
+
+    def get_chart(self, indicator):
+        # this is a serious hack for now
+        pie_class = 'sparkpie'
+        num, denom = self.get_indicator_data(indicator)
+        chart_template = (
+            '<span data-numerator="{num}" '
+            'data-denominator="{denom}" class="{pie_class}"></span>'
+        )
+        return format_html(chart_template, num=num,
+                           denom=denom - num,
+                           pie_class=pie_class)
+
 
 def flatten_config():
     props = """
